refactor: move time module to parser
