+++ crates/proof-of-sql-parser/Cargo.toml
@@ -15,9 +15,11 @@ doctest = true
 test = true
 
 [dependencies]
+arrow = { workspace = true }
 arrayvec = { workspace = true, features = ["serde"] }
 bigdecimal = { workspace = true }
 chrono = { workspace = true, features = ["serde"] }
+chrono-tz = { workspace = true, features = ["serde"] }
 lalrpop-util = { workspace = true, features = ["lexer", "unicode"] }
 serde = { workspace = true, features = ["serde_derive"] }
 thiserror = { workspace = true }
