+++ crates/proof-of-sql-parser/src/intermediate_time.rs
@@ -1,7 +1,39 @@
+use arrow::datatypes::TimeUnit as ArrowTimeUnit;
 use chrono::{offset::LocalResult, DateTime, TimeZone, Utc};
 use core::fmt;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
+
+/// Errors related to time operations, including timezone and timestamp conversions.
+#[derive(Error, Debug, Eq, PartialEq, Serialize, Deserialize)]
+pub enum TimeError {
+    /// Error indicating the timestamp unit provided is not supported.
+    #[error("unsupported timestamp unit: {0}")]
+    UnsupportedTimestampUnit(String),
+
+    /// Error when the timezone string provided cannot be parsed into a valid timezone.
+    #[error("invalid timezone string: {0}")]
+    InvalidTimezone(String),
+
+    /// Error indicating an invalid timezone offset was provided.
+    #[error("invalid timezone offset")]
+    InvalidTimezoneOffset,
+
+    /// Error when a timezone string fails to parse correctly.
+    #[error("failed to parse timezone string")]
+    TimeZoneStringParseError,
+
+    /// Error indicating a general failure to convert a timezone.
+    #[error("failed to convert timezone")]
+    TimeZoneConversionFailure,
+}
+
+// This exists because TryFrom<DataType> for ColumnType error is String
+impl From<TimeError> for String {
+    fn from(error: TimeError) -> Self {
+        error.to_string()
+    }
+}
 
 /// Errors from parsing a query string into an intermediate timestamp
 #[derive(Error, Debug, PartialEq, Eq)]
@@ -35,8 +67,8 @@ pub enum IntermediateTimestampError {
 }
 
 /// An intermediate type representing the time units from a parsed query
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-pub enum IntermediateTimeUnit {
+#[derive(Debug, Clone, Copy, Hash, Serialize, Deserialize, PartialEq, Eq)]
+pub enum PoSQLTimeUnit {
     /// Represents seconds with precision 0: ex "2024-06-20 12:34:56"
     Second,
     /// Represents milliseconds with precision 3: ex "2024-06-20 12:34:56.123"
@@ -47,13 +79,13 @@ pub enum IntermediateTimeUnit {
     Nanosecond,
 }
 
-impl fmt::Display for IntermediateTimeUnit {
+impl fmt::Display for PoSQLTimeUnit {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            IntermediateTimeUnit::Second => write!(f, "Second"),
-            IntermediateTimeUnit::Millisecond => write!(f, "Millisecond"),
-            IntermediateTimeUnit::Microsecond => write!(f, "Microsecond"),
-            IntermediateTimeUnit::Nanosecond => write!(f, "Nanosecond"),
+            PoSQLTimeUnit::Second => write!(f, "Second"),
+            PoSQLTimeUnit::Millisecond => write!(f, "Millisecond"),
+            PoSQLTimeUnit::Microsecond => write!(f, "Microsecond"),
+            PoSQLTimeUnit::Nanosecond => write!(f, "Nanosecond"),
         }
     }
 }
@@ -97,13 +129,13 @@ impl fmt::Display for IntermediateTimeZone {
 }
 
 /// Represents a fully parsed timestamp with detailed time unit and timezone information
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct IntermediateTimestamp {
     /// The datetime representation in UTC.
     pub timestamp: DateTime<Utc>,
 
     /// The precision of the datetime value, e.g., seconds, milliseconds.
-    pub timeunit: IntermediateTimeUnit,
+    pub timeunit: PoSQLTimeUnit,
 
     /// The timezone of the datetime, either UTC or a fixed offset from UTC.
     pub timezone: IntermediateTimeZone,
@@ -143,7 +175,7 @@ impl IntermediateTimestamp {
                 let offset_seconds = dt.offset().local_minus_utc();
                 IntermediateTimestamp {
                     timestamp: dt.with_timezone(&Utc),
-                    timeunit: IntermediateTimeUnit::Second,
+                    timeunit: PoSQLTimeUnit::Second,
                     timezone: IntermediateTimeZone::from_offset(offset_seconds),
                 }
             })
@@ -171,7 +203,7 @@ impl IntermediateTimestamp {
         match Utc.timestamp_opt(epoch, 0) {
             LocalResult::Single(timestamp) => Ok(IntermediateTimestamp {
                 timestamp,
-                timeunit: IntermediateTimeUnit::Second,
+                timeunit: PoSQLTimeUnit::Second,
                 timezone: IntermediateTimeZone::Utc,
             }),
             LocalResult::Ambiguous(_, _) => Err(IntermediateTimestampError::Ambiguous),
@@ -180,8 +212,177 @@ impl IntermediateTimestamp {
     }
 }
 
+impl From<PoSQLTimeUnit> for ArrowTimeUnit {
+    fn from(unit: PoSQLTimeUnit) -> Self {
+        match unit {
+            PoSQLTimeUnit::Second => ArrowTimeUnit::Second,
+            PoSQLTimeUnit::Millisecond => ArrowTimeUnit::Millisecond,
+            PoSQLTimeUnit::Microsecond => ArrowTimeUnit::Microsecond,
+            PoSQLTimeUnit::Nanosecond => ArrowTimeUnit::Nanosecond,
+        }
+    }
+}
+
+impl From<ArrowTimeUnit> for PoSQLTimeUnit {
+    fn from(unit: ArrowTimeUnit) -> Self {
+        match unit {
+            ArrowTimeUnit::Second => PoSQLTimeUnit::Second,
+            ArrowTimeUnit::Millisecond => PoSQLTimeUnit::Millisecond,
+            ArrowTimeUnit::Microsecond => PoSQLTimeUnit::Microsecond,
+            ArrowTimeUnit::Nanosecond => PoSQLTimeUnit::Nanosecond,
+        }
+    }
+}
+
+use chrono::FixedOffset;
+use chrono_tz::Tz;
+use std::{str::FromStr, sync::Arc};
+
+/// A typed TimeZone for a [`TimeStamp`]. It is optionally
+/// used to define a timezone other than UTC for a new TimeStamp.
+/// It exists as a wrapper around chrono-tz because chrono-tz does
+/// not implement uniform bit distribution
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash, Serialize, Deserialize)]
+pub struct PoSQLTimeZone(Tz);
+
+impl PoSQLTimeZone {
+    /// Convenience constant for the UTC timezone
+    pub const UTC: PoSQLTimeZone = PoSQLTimeZone(Tz::UTC);
+
+    /// Create a new ProofsTimeZone from a chrono TimeZone
+    pub fn new(tz: Tz) -> Self {
+        PoSQLTimeZone(tz)
+    }
+}
+
+impl TryFrom<IntermediateTimeZone> for PoSQLTimeZone {
+    type Error = TimeError;
+
+    fn try_from(tz: IntermediateTimeZone) -> Result<Self, Self::Error> {
+        match tz {
+            IntermediateTimeZone::Utc => Ok(PoSQLTimeZone::UTC),
+            IntermediateTimeZone::FixedOffset(seconds) => FixedOffset::east_opt(seconds)
+                .ok_or(TimeError::InvalidTimezoneOffset)
+                .and_then(|fixed_offset| {
+                    let datetime: DateTime<Utc> = Utc::now();
+                    let offset_datetime = datetime.with_timezone(&fixed_offset);
+                    let tz_string = offset_datetime.format("%Z").to_string();
+                    Tz::from_str(&tz_string)
+                        .map(PoSQLTimeZone)
+                        .map_err(|_| TimeError::TimeZoneStringParseError)
+                }),
+        }
+    }
+}
+
+impl From<&PoSQLTimeZone> for Arc<str> {
+    fn from(timezone: &PoSQLTimeZone) -> Self {
+        Arc::from(timezone.0.name())
+    }
+}
+
+impl From<Tz> for PoSQLTimeZone {
+    fn from(tz: Tz) -> Self {
+        PoSQLTimeZone(tz)
+    }
+}
+
+impl fmt::Display for PoSQLTimeZone {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl TryFrom<Option<Arc<str>>> for PoSQLTimeZone {
+    type Error = TimeError;
+
+    fn try_from(value: Option<Arc<str>>) -> Result<Self, Self::Error> {
+        match value {
+            Some(arc_str) => Tz::from_str(&arc_str)
+                .map(PoSQLTimeZone)
+                .map_err(|_| TimeError::InvalidTimezone("Invalid timezone string".to_string())),
+            None => Ok(PoSQLTimeZone(Tz::UTC)), // Default to UTC
+        }
+    }
+}
+
+impl TryFrom<&str> for PoSQLTimeZone {
+    type Error = &'static str;
+
+    fn try_from(value: &str) -> Result<Self, Self::Error> {
+        Tz::from_str(value)
+            .map(PoSQLTimeZone)
+            .map_err(|_| "Invalid timezone string")
+    }
+}
+
 #[cfg(test)]
 mod tests {
+    use super::*;
+    use chrono_tz::Tz;
+
+    #[test]
+    fn valid_timezones_convert_correctly() {
+        let valid_timezones = ["Europe/London", "America/New_York", "Asia/Tokyo", "UTC"];
+
+        for tz_str in &valid_timezones {
+            let arc_tz = Arc::new(tz_str.to_string());
+            // Convert Arc<String> to Arc<str> by dereferencing to &str then creating a new Arc
+            let arc_tz_str: Arc<str> = Arc::from(&**arc_tz);
+            let timezone = PoSQLTimeZone::try_from(Some(arc_tz_str));
+            assert!(timezone.is_ok(), "Timezone should be valid: {}", tz_str);
+            assert_eq!(
+                timezone.unwrap().0,
+                Tz::from_str(tz_str).unwrap(),
+                "Timezone mismatch for {}",
+                tz_str
+            );
+        }
+    }
+
+    #[test]
+    fn test_edge_timezone_strings() {
+        let edge_timezones = ["Etc/GMT+12", "Etc/GMT-14", "America/Argentina/Ushuaia"];
+        for tz_str in &edge_timezones {
+            let arc_tz = Arc::from(*tz_str);
+            let result = PoSQLTimeZone::try_from(Some(arc_tz));
+            assert!(result.is_ok(), "Edge timezone should be valid: {}", tz_str);
+            assert_eq!(
+                result.unwrap().0,
+                Tz::from_str(tz_str).unwrap(),
+                "Mismatch for edge timezone {}",
+                tz_str
+            );
+        }
+    }
+
+    #[test]
+    fn test_empty_timezone_string() {
+        let empty_tz = Arc::from("");
+        let result = PoSQLTimeZone::try_from(Some(empty_tz));
+        assert!(result.is_err(), "Empty timezone string should fail");
+    }
+
+    #[test]
+    fn test_unicode_timezone_strings() {
+        let unicode_tz = Arc::from("Europe/Paris\u{00A0}"); // Non-breaking space character
+        let result = PoSQLTimeZone::try_from(Some(unicode_tz));
+        assert!(
+            result.is_err(),
+            "Unicode characters should not be valid in timezone strings"
+        );
+    }
+
+    #[test]
+    fn test_null_option() {
+        let result = PoSQLTimeZone::try_from(None);
+        assert!(result.is_ok(), "None should convert without error");
+        assert_eq!(result.unwrap().0, Tz::UTC, "None should default to UTC");
+    }
+}
+
+#[cfg(test)]
+mod timezone_tests {
     use super::*;
     use chrono::TimeZone;
 
@@ -229,7 +430,7 @@ mod tests {
     fn test_unix_epoch_timestamp_parsing() {
         let unix_time = 1_593_000_000; // Example Unix timestamp (seconds since epoch)
         let expected_datetime = Utc.timestamp_opt(unix_time, 0).unwrap();
-        let expected_unit = IntermediateTimeUnit::Second; // Assuming basic second precision for Unix timestamp
+        let expected_unit = PoSQLTimeUnit::Second; // Assuming basic second precision for Unix timestamp
         let input = unix_time; // Simulate input as string since Unix times are often transmitted as strings
         let result = IntermediateTimestamp::to_timestamp(input).unwrap();
 
@@ -279,7 +480,7 @@ mod tests {
     fn test_timestamp_with_seconds() {
         let input = "2023-06-26T12:34:56Z";
         let expected_time = Utc.with_ymd_and_hms(2023, 6, 26, 12, 34, 56).unwrap();
-        let expected_unit = IntermediateTimeUnit::Second;
+        let expected_unit = PoSQLTimeUnit::Second;
         let result = IntermediateTimestamp::try_from(input).unwrap();
         assert_eq!(result.timestamp, expected_time);
         assert_eq!(result.timeunit, expected_unit);
