+++ crates/proof-of-sql/src/base/database/owned_table_utility.rs
@@ -14,12 +14,12 @@
 //! ]);
 //! ```
 use super::{OwnedColumn, OwnedTable};
-use crate::base::{
-    scalar::Scalar,
-    time::{timestamp::PoSQLTimeUnit, timezone::PoSQLTimeZone},
-};
+use crate::base::scalar::Scalar;
 use core::ops::Deref;
-use proof_of_sql_parser::Identifier;
+use proof_of_sql_parser::{
+    intermediate_time::{PoSQLTimeUnit, PoSQLTimeZone},
+    Identifier,
+};
 
 /// Creates an OwnedTable from a list of (Identifier, OwnedColumn) pairs.
 /// This is a convenience wrapper around OwnedTable::try_from_iter primarily for use in tests and
@@ -212,7 +212,8 @@ pub fn decimal75<S: Scalar>(
 /// ```
 /// use proof_of_sql::base::{database::owned_table_utility::*,
 ///     scalar::Curve25519Scalar,
-///     time::{timestamp::PoSQLTimeUnit, timezone::PoSQLTimeZone}
+/// };
+/// use proof_of_sql_parser::{intermediate_time::{PoSQLTimeUnit, PoSQLTimeZone},
 /// };
 /// use chrono_tz::Europe::London;
 ///
