+++ crates/proof-of-sql/src/base/mod.rs
@@ -11,5 +11,3 @@ pub mod scalar;
 mod serialize;
 pub(crate) use serialize::{impl_serde_for_ark_serde_checked, impl_serde_for_ark_serde_unchecked};
 pub(crate) mod slice_ops;
-/// Stores all functionality relelvant to time
-pub mod time;
