--- crates/proof-of-sql/src/base/time/mod.rs
@@ -1,6 +1,0 @@
-/// Error module to store all errors related to timestamp parsing
-pub mod error;
-/// Stores all functionality relelvant to timestamps and time units
-pub mod timestamp;
-/// Stores functionality relevant to timezones
-pub mod timezone;
