+++ crates/proof-of-sql/src/sql/parse/error.rs
@@ -1,5 +1,5 @@
-use crate::base::{database::ColumnType, math::decimal::DecimalError, time::error::TimeError};
-use proof_of_sql_parser::{Identifier, ResourceId};
+use crate::base::{database::ColumnType, math::decimal::DecimalError};
+use proof_of_sql_parser::{intermediate_time::TimeError, Identifier, ResourceId};
 use thiserror::Error;
 
 /// Errors from converting an intermediate AST into a provable AST.
